NodeClass: use long offsets, extract access to Node and NodeList members into helper methods

+++ com.oracle.max.vm.ext.c1xgraal/src/com/oracle/max/vm/ext/c1xgraal/C1XGraal.java
@@ -152,7 +152,7 @@ public class C1XGraal implements RuntimeCompiler {
         if (isHosted() && phase == Phase.SERIALIZING_IMAGE) {
             NodeClass.rescanAllFieldOffsets(new CalcOffset() {
                 @Override
-                public int getOffset(Field field) {
+                public long getOffset(Field field) {
                     return FieldActor.fromJava(field).offset();
                 }
             });
